# elderly-care-backend/docs/swagger.yaml
definitions:
  dto.AcceptSOSRequest:
    properties:
      volunteer_id:
        description: 改为uint
        type: integer
    required:
    - volunteer_id
    type: object
  dto.AcceptTaskRequest:
    properties:
      volunteer_id:
        description: 改为uint
        type: integer
      volunteer_lat:
        type: number
      volunteer_lng:
        type: number
    required:
    - volunteer_id
    - volunteer_lat
    - volunteer_lng
    type: object
  dto.CreateTaskRequest:
    properties:
      address:
        type: string
      category:
        type: string
      creator_id:
        description: 改为uint
        type: integer
      deadline:
        type: string
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      reward:
        type: number
      title:
        type: string
    required:
    - address
    - category
    - creator_id
    - description
    - latitude
    - longitude
    - title
    type: object
  dto.ResolveSOSRequest:
    properties:
      resolved_by:
        description: 改为uint
        type: integer
    required:
    - resolved_by
    type: object
  dto.SOSRequest:
    properties:
      address:
        type: string
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      severity:
        type: string
      user_id:
        description: 改为uint
        type: integer
    required:
    - address
    - description
    - latitude
    - longitude
    - user_id
    type: object
...
  vo.ResponseVO:
    properties:
      data: {}
      msg:
        type: string
      success:
        type: boolean
    type: object
  vo.SOSRecordVO:
    properties:
      address:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      severity:
        type: string
      status:
        type: string
      task_id:
        type: integer
      timeout:
        type: integer
      user_id:
        type: integer
    type: object
  vo.SOSResponseVO:
    properties:
      matches:
        items:
          additionalProperties: true
          type: object
        type: array
      sos_id:
        type: integer
      task_id:
        type: integer
      timeout:
        type: integer
    type: object
  vo.TaskMatchVO:
    properties:
      matches:
        items:
          additionalProperties: true
          type: object
        type: array
      task:
        $ref: '#/definitions/vo.TaskVO'
    type: object
  vo.TaskVO:
    properties:
      address:
        type: string
      category:
        type: string
      created_at:
        type: string
      creator_id:
        type: integer
      creator_name:
        type: string
      description:
        type: string
      distance:
        type: number
      id:
        type: integer
      is_emergency:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      reward:
        type: number
      status:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
  description: 养老服务平台接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: elderly-care-backend
  version: "1.0"
paths:
...
  /sos/{sosId}/accept:
    post:
      consumes:
      - application/json
      description: 志愿者接受SOS紧急求助
      parameters:
      - description: SOS记录ID
        in: path
        name: sosId
        required: true
        type: integer
      - description: 接受SOS请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AcceptSOSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 接受SOS求助
      tags:
      - SOS模块
  /sos/{sosId}/resolve:
    put:
      consumes:
      - application/json
      description: 标记SOS紧急求助为已解决
      parameters:
      - description: SOS记录ID
        in: path
        name: sosId
        required: true
        type: integer
      - description: 解决SOS请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResolveSOSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 解决SOS求助
      tags:
      - SOS模块
  /sos/current:
    get:
      consumes:
      - application/json
      description: 获取用户当前的SOS求助状态
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.ResponseVO'
            - properties:
                data:
                  $ref: '#/definitions/vo.SOSRecordVO'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前SOS状态
      tags:
      - SOS模块
  /sos/emergency:
    post:
      consumes:
      - application/json
      description: 用户触发SOS紧急求助
      parameters:
      - description: SOS求助请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SOSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.ResponseVO'
            - properties:
                data:
                  $ref: '#/definitions/vo.SOSResponseVO'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 触发紧急求助
      tags:
      - SOS模块
  /tasks:
    post:
      consumes:
      - application/json
      description: 创建新的求助任务
      parameters:
      - description: 创建任务请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.ResponseVO'
            - properties:
                data:
                  $ref: '#/definitions/vo.TaskMatchVO'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建任务
      tags:
      - 任务模块
  /tasks/{taskId}/accept:
    post:
      consumes:
      - application/json
      description: 志愿者接受任务
      parameters:
      - description: 任务ID
        in: path
        name: taskId
        required: true
        type: integer
      - description: 接受任务请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AcceptTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.ResponseVO'
      security:
      - ApiKeyAuth: []
      summary: 接受任务
      tags:
      - 任务模块
  /tasks/nearby:
    get:
      consumes:
      - application/json
      description: 根据位置获取附近的任务列表
      parameters:
      - description: 纬度
        in: query
        name: lat
        required: true
        type: number
      - description: 经度
        in: query
        name: lng
        required: true
        type: number
      - default: 5000
        description: 搜索半径(米)
        in: query
        name: radius
        type: number
      - description: 任务分类
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/vo.ResponseVO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/vo.TaskVO'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取附近任务
      tags:
      - 任务模块
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"